LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/key-id-back-cross-document-expected.txt

CONSOLE MESSAGE: Error: assert_equals: expected 1 but got 0


Harness Error (TIMEOUT), message = null

TIMEOUT NavigationHistoryEntry's key and id on cross-document back navigation Test timed out
